Renames node package to host. Pulls host interface out into common package.
