+++ go/common/host/host.go
