+++ go/host/rpc/clientapi/client_api_obscuroscan.go
@@ -4,10 +4,11 @@ import (
 	"fmt"
 	"math/big"
 
+	"github.com/obscuronet/go-obscuro/go/common/host"
+
 	gethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/obscuronet/go-obscuro/go/common"
-	"github.com/obscuronet/go-obscuro/go/host"
 )
 
 // ObscuroScanAPI implements ObscuroScan-specific JSON RPC operations.
